requirements.txt
torch
diffusers
customtkinter
python-dotenv
Pillow
